Made improvements to the Schedule and Thermostat services and controllers

+++ src/main/java/io/github/incplusplus/betterstat/service/ScheduleServiceImpl.java
@@ -2,6 +2,7 @@ package io.github.incplusplus.betterstat.service;
 
 import io.github.incplusplus.betterstat.persistence.model.Schedule;
 import io.github.incplusplus.betterstat.persistence.repositories.ScheduleRepository;
+import io.github.incplusplus.betterstat.web.exception.ObjectNotFoundException;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,6 +25,16 @@ public class ScheduleServiceImpl implements ScheduleService {
   }
 
   @Override
+  public Schedule updateSchedule(String id, Schedule schedule) throws ObjectNotFoundException {
+    if (!this.existsById(id)) throw new ObjectNotFoundException(id, Schedule.class);
+    // If it already exists, set the ID of the DTO to the ID we know exists in our repository.
+    schedule.setId(id);
+    // Saving an object with an ID that already exists will simply update it.
+    // https://stackoverflow.com/a/56207430/1687436
+    return scheduleRepository.save(schedule);
+  }
+
+  @Override
   public Optional<Schedule> getScheduleById(String id) {
     return scheduleRepository.findById(id);
   }
@@ -39,7 +50,12 @@ public class ScheduleServiceImpl implements ScheduleService {
   }
 
   @Override
-  public void deleteById(String id) {
+  public Schedule deleteById(String id) throws ObjectNotFoundException {
+    Optional<Schedule> scheduleOptional = this.getScheduleById(id);
+    if (scheduleOptional.isEmpty()) {
+      throw new ObjectNotFoundException(id, Schedule.class);
+    }
     scheduleRepository.deleteById(id);
+    return scheduleOptional.get();
   }
 }
